back-end/app/requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
boto==2.49.0
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
enum34==1.1.6
firebase-admin==2.16.0
Flask==1.0.2
futures==3.2.0
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.32.1
google-cloud-storage==1.15.0
google-compute-engine==2.8.13
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.20.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.24.2
Werkzeug==0.15.2
flask-cors==3.0.7
flask-mail==0.9.1
requests-toolbelt==0.9.1
